+++ NEWS.md
@@ -1,3 +1,9 @@
+# TreeTools 1.4.2
+
+ - Support star trees in `RootTree()`.
+ - Improve memory handling in `root_on_node()`.
+ - Documentation linkage.
+
 # TreeTools 1.4.1
 
  - `MSTEdges()` supports distance matrices with > 256 entries.